[refactor] migrate auto_charge_neutralisation to autofunctions/charge_neutralisation.py
